Added intermediate assignment tests
